+++ feos-core/src/python/phase_equilibria.rs
@@ -463,11 +463,25 @@ macro_rules! impl_phase_equilibrium {
         }
 
         /// Phase diagram for a pure component or a binary mixture.
+        ///
+        /// Parameters
+        /// ----------
+        /// phase_equilibria : [PhaseEquilibrium]
+        ///     A list of individual phase equilibria.
+        ///
+        /// Returns
+        /// -------
+        /// PhaseDiagram : the resulting phase diagram
         #[pyclass(name = "PhaseDiagram")]
-        pub struct PyPhaseDiagram(PhaseDiagram<SIUnit, $eos>);
+        pub struct PyPhaseDiagram(PhaseDiagram<SIUnit, $eos, 2>);
 
         #[pymethods]
         impl PyPhaseDiagram {
+            #[new]
+            fn new(phase_equilibria: Vec<PyPhaseEquilibrium>) -> Self {
+                Self(PhaseDiagram::new(phase_equilibria.into_iter().map(|p| p.0).collect()))
+            }
+
             /// Calculate a pure component phase diagram.
             ///
             /// Parameters